my_tide_app/requirements.txt
# my_tide_app/requirements.txt

Flask==3.0.3
pandas==2.2.2
requests==2.32.3
pytz==2024.1
geopy==2.4.1
python-dotenv==1.0.1
gunicorn==22.0.0
